refactor: use io.qnop as the base Java package
Rename the base package from de.qnop to io.qnop across all modules
(directories, package-info, the ArchUnit test) and align the Gradle group
coordinate and docs accordingly.

Refs #1
