--- qnop-app/src/main/java/de/qnop/app/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Composition root / Spring Boot bootstrap (Community build). */
-package de.qnop.app;
