+++ qnop-app/src/main/java/io/qnop/app/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Composition root / Spring Boot bootstrap (Community build). */
+package io.qnop.app;
