--- qnop-application/src/main/java/de/qnop/application/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Use cases and ports (repository and provider interfaces). */
-package de.qnop.application;
