+++ qnop-application/src/main/java/io/qnop/application/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Use cases and ports (repository and provider interfaces). */
+package io.qnop.application;
