--- qnop-document/src/main/java/de/qnop/document/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Text extraction, conversion and annotation anchoring adapters. */
-package de.qnop.document;
