+++ qnop-document/src/main/java/io/qnop/document/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Text extraction, conversion and annotation anchoring adapters. */
+package io.qnop.document;
