--- qnop-domain/src/main/java/de/qnop/domain/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Framework-free entities, value objects and the review workflow state machine. */
-package de.qnop.domain;
