+++ qnop-domain/src/main/java/io/qnop/domain/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Framework-free entities, value objects and the review workflow state machine. */
+package io.qnop.domain;
