--- qnop-persistence/src/main/java/de/qnop/persistence/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** JPA adapters and Flyway migrations (PostgreSQL). */
-package de.qnop.persistence;
