+++ qnop-persistence/src/main/java/io/qnop/persistence/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** JPA adapters and Flyway migrations (PostgreSQL). */
+package io.qnop.persistence;
