--- qnop-security/src/main/java/de/qnop/security/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Authentication, authorization and the user and team model. */
-package de.qnop.security;
