+++ qnop-security/src/main/java/io/qnop/security/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Authentication, authorization and the user and team model. */
+package io.qnop.security;
