--- qnop-spi/src/main/java/de/qnop/spi/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** Extension-point interfaces and DTOs (the AGPL/commercial boundary). */
-package de.qnop.spi;
