+++ qnop-spi/src/main/java/io/qnop/spi/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** Extension-point interfaces and DTOs (the AGPL/commercial boundary). */
+package io.qnop.spi;
