--- qnop-storage/src/main/java/de/qnop/storage/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** StorageProvider adapter over the S3 API. */
-package de.qnop.storage;
