+++ qnop-storage/src/main/java/io/qnop/storage/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** StorageProvider adapter over the S3 API. */
+package io.qnop.storage;
