--- qnop-web/src/main/java/de/qnop/web/package-info.java
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-only
-
-/** REST controllers, DTO mapping and OpenAPI. */
-package de.qnop.web;
