+++ qnop-web/src/main/java/io/qnop/web/package-info.java
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-only
+
+/** REST controllers, DTO mapping and OpenAPI. */
+package io.qnop.web;
